requirements.txt
Django==2.1.3
django-cors-headers==2.4.0
django-filter==2.0.0
django-redis==4.9.0
django-rest-swagger==2.2.0
djangorestframework==3.9.0
psycopg2-binary==2.7.6.1
